Add a method for Tutorial to find its course
Previously, there was no simple way to find the course that matches a
specific Tutorial when trying to create a Tutorial object.

This is a common scenario, so the classes had to manually search through
a set of Courses to find a Course that would be pointed at by the
Tutorial. This is very annoying, highly inefficient, and requires much
code repetition.

Let's add a new method to Tutorial that would take a set of Courses,
search through it, and then return the relevant Course. This would
result in much less repetitive code.