Change AddressBooKParser's parseCommand method
AddressBookParser used to use a very lengthy switch/case statement to
decide which command word was said.

This can be shortened by using a map of Strings to CheckedFunctions
(functional interfaces that behave like Functions but can throw an
error). While this doesn't eliminate the complexity of checking, it
does allow for autocomplete to be implemented easier with a unified
list of command words, with not much additional cost to current
implementations of parsing methods.

Let's change it to this equally complex but more easily extensible
method of implementation!